+++ cities/rio-de-janeiro-theme/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2026.1.2
+
+- Published to the Open VSX Registry.
+
 ## 2026.1.1
 
 - Initial release.