+++ nature/autumn-theme/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2026.1.3
+
+- Published to the Open VSX Registry.
+
 ## 2026.1.2
 
 - Fix README.md.